refacto(SearchableCompaniesList) : use same component as in <CompaniesList /> to handle search
